cuda: FLIP heads to bf16 -- the two lines increments 1-7 were built for (L033)
    PULSAR_HEADS_ELT_SIZE  4u -> 2u
    pulsar_heads_t         float -> __nv_bfloat16

Everything else already moved.  Increments 1-7 converted the whole plumbing
inert at f32 -- both byte-strided row views, both attention producers,
rope_tail (templated, its f32 arm SASS-identical), the grouped "a" projection
and its three kernels, the 8 attention row pointers and 12 stores, the 15 entry
casts, and the three diagnostics that had to learn to WIDEN rather than
reinterpret.  The full 18-gate suite went GREEN on that at ad88cf6, byte-exact,
so everything below this line is attributable to the width and nothing else.

⚠ THIS ONE IS NOT A NO-OP, AND THE BYTE GATE IS EXPECTED TO FAIL.  That is the
point: under the source standard a byte-gate FAIL is a question for
cuda-reference-gate, not a revert (memory: bit-exact-measures-us-not-source).
The direction was already measured when the plumbing was designed, on CONFIDENT
depths only:

    f32 -> f16  : +7.23% FURTHER from source
    f32 -> bf16 : -29.67% CLOSER

They point OPPOSITE ways, and the old all-depths view called both "neutral"
because two flat rows masked the difference.  bf16 is the source's own residual
dtype (ds4-source-numerics); f16 is a dtype the source has nowhere.  So this is
a FIDELITY IMPROVEMENT, not a tradeoff -- which is why it lands even though:

⚠ IT IS NOT A SPEED CHANGE, and must not be sold as one.  The engine is
latency-bound; attn_f16 sits at ~22% memory throughput and is stalled on its KV
LOADS, so halving a STORE cannot relieve it.  Today's L100 result sharpens this
further -- fewer/wider memory traffic did nothing on the emptiest kernel in the
engine (wider-loads-are-not-mlp-2026-08-24.md).  Expect ~0.  The wins are
fidelity and a 512 -> 256 MiB allocation.

Compile-checked under the flip before committing: all four heads-touching TUs
clean, and grouped_fp8mx_a_warp8_a8_kernel unchanged at 56 registers, 0 spills
(it consumes quantised E4M3 activations, not heads).

Still to measure, and the results decide whether this stays:
  1. cuda-reference-gate direction on confident depths -- the actual verdict;
  2. decode A/B -- every format narrowing gets one (memory: bit-exact-is-not-free,
     where a bit-exact FP4 narrowing cost 1.9% decode and passed all 18 gates);
  3. the rest of the suite, which should be unaffected.

+++ src/cuda/pulsar_cuda_internal.h
@@ -274,11 +274,12 @@ static_assert(sizeof(pulsar_hc_t) == PULSAR_HC_ELT_SIZE, "pulsar_hc_t size must 
  * PULSAR_HEADS_ELT_SIZE in pulsar_gpu.h.  Same contract as pulsar_hc_t: the
  * STORAGE narrows, every kernel still loads to f32 and accumulates in f32.
  *
- * Currently float -- the plumbing lands before the flip (see the macro's note).
- * The static_assert below is the whole point of declaring the two together: the
+ * FLIPPED to __nv_bfloat16 2026-08-24, after increments 1-7 landed the whole
+ * plumbing inert at f32 and the full suite went green on it byte-exact.  The
+ * static_assert below is the whole point of declaring the two together: the
  * Q buffer's width mismatch shipped as a real defect because changing either
  * side alone compiled clean. */
-typedef float pulsar_heads_t;
+typedef __nv_bfloat16 pulsar_heads_t;
 /* Both casts are valid for float AND __nv_bfloat16 (the bf16 type carries a
  * float conversion operator and a float constructor), so the flip really is
  * one line in pulsar_gpu.h plus one in the typedef above. */
